+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Fixed bug that saved contacts from groups came without number in chatwoot
 * Fixed problem to receive csat in chatwoot
 * Fixed require fileName for document only in base64 for send media message
+* Bug fix when sending mobile message change contact name to number in chatwoot
 
 # 1.2.2 (2023-07-15 09:36)