+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added restart instance endpoint
 * Created automation for creating instances in the chatwoot bot with the command '#inbox_whatsapp:<INSTANCE_NAME>'
 * Change Baileys version to: 6.4.0
+* Send contact in chatwoot
 
 ### Fixed